non finished auth

+++ db/filc.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS `user_types` (
 );
 
 CREATE TABLE IF NOT EXISTS `users` (
-    `id` BIGINT NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id` INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `type_id` TINYINT NOT NULL,
     `name` VARCHAR(255) NOT NULL,
     `email` VARCHAR(255) NOT NULL UNIQUE,
@@ -16,16 +16,16 @@ CREATE TABLE IF NOT EXISTS `users` (
 );
 
 CREATE TABLE IF NOT EXISTS `courses` (
-    `id` BIGINT NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    `teacher_id` BIGINT NOT NULL,
+    `id` INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `teacher_id` INT NOT NULL,
     `name` VARCHAR(255) NOT NULL,
 
     FOREIGN KEY (`teacher_id`) REFERENCES `users`(`id`) ON DELETE CASCADE
 );
 
 CREATE TABLE IF NOT EXISTS `student_courses` (
-    `user_id` BIGINT NOT NULL,
-    `course_id` BIGINT NOT NULL,
+    `user_id` INT NOT NULL,
+    `course_id` INT NOT NULL,
 
     FOREIGN KEY (`user_id`) REFERENCES `users`(`id`) ON DELETE CASCADE,
     FOREIGN KEY (`course_id`) REFERENCES `courses`(`id`) ON DELETE CASCADE
@@ -38,9 +38,9 @@ CREATE TABLE IF NOT EXISTS `post_types` (
 );
 
 CREATE TABLE IF NOT EXISTS `posts` (
-    `id` BIGINT NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id` INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `type_id` TINYINT NOT NULL,
-    `course_id` BIGINT NOT NULL,
+    `course_id` INT NOT NULL,
     `content` JSON NOT NULL,
 
     FOREIGN KEY (`course_id`) REFERENCES `courses`(`id`) ON DELETE CASCADE,
@@ -48,8 +48,8 @@ CREATE TABLE IF NOT EXISTS `posts` (
 );
 
 CREATE TABLE IF NOT EXISTS `grades` (
-    `id` BIGINT NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    `course_id` BIGINT NOT NULL,
+    `id` INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `course_id` INT NOT NULL,
     `weight` DOUBLE NOT NULL DEFAULT 1,
     `name` VARCHAR(100),
 
@@ -57,8 +57,8 @@ CREATE TABLE IF NOT EXISTS `grades` (
 );
 
 CREATE TABLE IF NOT EXISTS `student_grades` (
-    `student_id` BIGINT NOT NULL,
-    `grade_id` BIGINT NOT NULL,
+    `student_id` INT NOT NULL,
+    `grade_id` INT NOT NULL,
     `grade` tinyint CHECK (grade BETWEEN 1 AND 5 OR grade IS NULL),
 
     FOREIGN KEY (`student_id`) REFERENCES `users`(`id`) ON DELETE CASCADE,
@@ -67,7 +67,7 @@ CREATE TABLE IF NOT EXISTS `student_grades` (
 
 CREATE TABLE IF NOT EXISTS `sessions` (
     `id` VARCHAR(64) NOT NULL PRIMARY KEY,
-    `user_id` BIGINT NOT NULL,
+    `user_id` INT NOT NULL,
     `expires_at` DATETIME NOT NULL,
 
     FOREIGN KEY (`user_id`) REFERENCES `users`(`id`) ON DELETE CASCADE
